+++ packages/icon/src/icons/ChevronDownIcon.tsx
@@ -4,7 +4,6 @@ import { createIcon } from '../Icon';
 export const ChevronDownIcon = createIcon(
 	<path
 		d="M21 8L12.3536 16.6464C12.1583 16.8417 11.8417 16.8417 11.6464 16.6464L3 8"
-		stroke="#313131"
 		strokeWidth="2"
 		strokeLinecap="round"
 		strokeLinejoin="round"
